refactor: Rename `helper` to `useShouldShowDefaultRedirectionSnackbar`
